+++ src/molassembler/lib/molassembler/Graph/Canonicalization.h
@@ -1,6 +1,21 @@
-/* @file
+/*! @file
+ * @brief Provides canonicalization of Molecule instances
  * @copyright ETH Zurich, Laboratory for Physical Chemistry, Reiher Group.
  *   See LICENSE.txt for details.
+ *
+ *
+ * Graph canonicalization is an algorithm that generates an isomorphic graph
+ * to its input. The important property is that the algorithm's output graph is
+ * identical for all graphs that are isomorphic to its input. For two canonical
+ * graphs, an isomorphism test reduces to a fast identity test.
+ *
+ * Naturally, graph canonicalization is computationally at least as hard as an
+ * isomorphism test on non-canonical graphs, so little time (if any) is saved
+ * by canonicalizing two molecules separately, then using the identity
+ * comparison. However, it is worth canonicalizing molecules when it is to be
+ * expected that repeated isomorphism computations need to be performed upon
+ * many molecules. For instance, checking whether a molecule is already present
+ * in a set of molecules can be sped up considerably.
  */
 
 #ifndef INCLUDE_MOLASSEMBLER_GRAPH_CANONICALIZATION_H
@@ -32,6 +47,10 @@ using WideHashType = boost::multiprecision::uint128_t;
  * @throws std::invalid_argument If vector of hashes length does not match
  *   the molecule's number of vertices.
  *
+ * @warning The canonical form of the molecule is influenced by the provided
+ *   hashes. Any comparisons on the canonical form must use the same comparison
+ *   bitmask as when generating the hashes for this function call here.
+ *
  * @return The canonical labeling sequence of atom indices into @p mol.
  */
 std::vector<int> canonicalAutomorphism(
@@ -46,6 +65,8 @@ std::vector<int> canonicalAutomorphism(
  * @param firstHashes Vertex coloring hashes for the first molecule
  * @param secondGraph The graph of the second molecule to compare
  * @param secondHashes Vertex coloring hashes for the second molecule
+ *
+ * @note A word on complexity: Incurs four graph copies, two each.
  *
  * @return Whether the two graphs are isomorphic
  */